RUB-292 go: split consensus helper shape
